Started adding support for mis-aligned load values and messages. Some alignment scenarios may require additional Spike support. Not fully tested

+++ pk/pk.h
@@ -82,7 +82,7 @@ typedef struct {
 typedef struct {
     word_t candidate_messages[64]; //Support UP TO 64 candidate messages
     size_t size;
-    size_t load_message_offset;
+    long load_message_offset;
     size_t load_size;
 } due_candidates_t;
 
@@ -108,7 +108,7 @@ int copy_cacheline(due_cacheline_t* dest, due_cacheline_t* src); //MWG
 int copy_candidates(due_candidates_t* dest, due_candidates_t* src); //MWG
 int copy_trapframe(trapframe_t* dest, trapframe_t* src); //MWG
 unsigned long decode_rd(long insn); //MWG
-int writeback_recovered_message(word_t* recovered_message, unsigned long load_size, unsigned long offset, trapframe_t* tf); //MWG
+int writeback_recovered_message(word_t* recovered_message, unsigned long load_size, long offset, trapframe_t* tf); //MWG
 
 typedef struct {
   int elf64;
